+++ src/OpenCL/kernels/init.cl
@@ -1,4 +1,4 @@
-__kernel void init(__global double* vec, double value, int size){
+__kernel void init(__global float* vec, float value, int size){
 	// int id = blockIdx.x * blockDim.x + threadIdx.x;
     int id = get_global_id(0);
 	if(id < size) vec[id] = value;
